Added beforeSubmit step in DetailQuestion
